moving hibachi decorators logic into it's own class
